# doc/api.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: WASA Photo REST API
  description: |
    This is the API specification for the WASA Photo REST API.

    It's part of the WASA Photo project for the Web and Software Architecture course at Sapienza University of Rome.

    The project description and details are availabe at [this link](http://gamificationlab.uniroma1.it/notes/Project.pdf)
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    bearerAuth:
      description: |
        This security scheme is used to authenticate a user.
        The user must be signed in to perform the request.
        The user must pass the bearer token in the Authorization header.
      type: http
      scheme: bearer
  schemas:
    username:
      description: Username of a user
      type: string
      minLength: 4
      maxLength: 20
      pattern: '^[a-zA-Z0-9_-]+$'
      readOnly: true
      example: mariorossi99
    resourceId:
      description: The Universally Unique Identifier (follows the standard)
      type: string
      format: uuid
      example: 550e8400-e29b-41d4-a716-446655440000
      readOnly: true
    date:
      description: representation of Date and time
      type: string
      format: date-time
      example: 2023-01-01T00:00:00Z
      readOnly: true
    caption:
      description: A caption for a post or a comment
      type: string
      minLength: 1
      maxLength: 800
      pattern: '.*'
      example: "This is a caption"
      readOnly: false
    image:
      description: An uri to an image
      type: string
      format: uri
      example: "localhost:8080/users/mariorossi99/posts/1/image"
      readOnly: false
    counter:
      description: A simple counter,
                    keeps track of the number of followers, likes, etc.
      type: integer
      example: 8
      readOnly: true
    userCollection:
      description: A list of usersnames, 
                   it is used to keep track of the users that liked a post or a comment
      type: array
      minLength: 0
      maxLength: 69420
      items:
        anyOf:
          - $ref: '#/components/schemas/User/properties/username'
          - $ref: '#/components/schemas/User/properties/userId'
          - $ref: '#/components/schemas/User/properties/profileImage'
      readOnly: false
...
    UserPost:
      title: UserPost
      type: object
      description: A post created by a user
      properties:
        postId:
          $ref: '#/components/schemas/resourceId'
        authorUsername:
          $ref: '#/components/schemas/username'
        authorId:
          $ref: '#/components/schemas/User/properties/userId'
        creationDate:
          $ref: '#/components/schemas/date'
        caption:
          $ref: '#/components/schemas/caption'
        image:
          $ref: '#/components/schemas/image'
        likeCount:
          $ref: '#/components/schemas/counter'
        commentCount:
          $ref: '#/components/schemas/counter'
      required:
        - postId
        - authorUsername
        - creationDate
        - caption
...
    postStream:
      description: An array of posts, 
                   can either be the list of posts of a user or the list of posts of the users followed by a user 
      type: array
      minLength: 0
      maxLength: 69420
      items:
        $ref: '#/components/schemas/UserPost/properties/postId'
...
    Error:
      description: An error message
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]$'
      required:
        - message
    Success:
      description: A successful request
      type: object
      properties:
        message:
          type: string
          minLength: 1
          maxLength: 100
          pattern: '^[a-zA-Z0-9_-]$'
        body:
          type: object
      required:
        - message
  parameters:
    userId:
      name: userId
      in: path
      description: The userId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
    postId:
      name: postId
      in: path
      description: The postId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
    commentId:
      name: commentId
      in: path
      description: The commentId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
...
    likeId:
      name: likeId
      in: path
      description: The likeId that is being requested
      required: true
      schema:
        $ref: '#/components/schemas/resourceId'
  responses:
...
    NotFound: #for 404
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized: #for 401
      description: The request requires user authentication, 
                    or the user is not authorized to perform the requested action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    BadRequest: #for 400
      description: The request was not valid, 
                    the request body is missing or malformed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError: #for 500
      description: An error occurred on the server side
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Created: #for 201
      description: The resource was successfully created
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/resourceId'
    Ok: #for 200
      description: The request has succeeded
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Success'
...
    UserPost: # 200 response for a post 
      description: A post created by a user
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserPost'
...
  requestBodies:
    UserPost:
      description: A post created by a user
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/UserPost'
...
paths:
...
  /users/{userId}/posts:
    description: This endpoint handles the collection of posts of a user.
    parameters:
      - $ref: '#/components/parameters/userId'
  
    post: 
      tags: ["post"]
      operationId: createPost
      summary: Create a new post
      description: |
        This endpoint is used to create a new post.
        The userId is passed as a path parameter.
        The post details are passed in the request body.
        The response will retun the id of the new post.
      requestBody:
        $ref: '#/components/requestBodies/UserPost'
      responses:
        "201":
          $ref: '#/components/responses/Created'
        "400": #the request body is missing or malformed
          $ref: '#/components/responses/BadRequest'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
    
    get: #get user posts
      tags: ["post"]
      operationId: getUserPosts
      summary: Get all the posts of a user
      description: |
        This endpoint is used to get all the posts of a user.
        The userId is passed as a path parameter.
        The response will retun the list of posts of the user.
      responses:
        "200":
          description: The list of posts of the user
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/postStream'
        "404": #user not found
          $ref: '#/components/responses/NotFound'
  /users/{userId}/posts/{postId}:
    description: This endpoint handles a single post of a user.
                  it can be used to get the details of a post, delete it, or update it.
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/postId'
    
    get:
      tags: ["post"]
      operationId: getPost
      summary: Get the details of a post
      description: |
        This request is used to get the details of a post.
        The userId and the postId are passed as path parameters.
        The response will retun the details of the post in the body.
      responses:
        "200":
          $ref: '#/components/responses/UserPost'
        "404": #post not found
          $ref: '#/components/responses/NotFound'
        "401":
          $ref: '#/components/responses/Unauthorized'
      
    put:
      tags: ["post"]
      operationId: editPost
      summary: Edit a post
      description: |
        This request is used to edit a post.
        The userId and the postId are passed as path parameters.
        The new post details are passed in the request body.
        The response will retun the id of the new post.
      requestBody:
        $ref: '#/components/requestBodies/UserPost'
      responses:
        "202": #update successful
          $ref: '#/components/responses/Ok'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
        "401":
          $ref: '#/components/responses/Unauthorized'

    delete:
      tags: ["post"]
      operationId: deletePost
      summary: Delete the post
      description: Deletes the given post
      responses:
        "200": #update successful
          $ref: '#/components/responses/Ok'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/posts/{postId}/likes:
    description: This endpoint handles the collection of likes of a post.
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/postId'

    get:
      tags: ["like"]
      operationId: getPostLikes
      summary: Get all the likes of a post
      description: |
        This request is used to get all the users that liked a post.
        The postId is passed as a path parameter.
        The response will retun the list of likes of the post.
      responses:
        "200":
          description: The list of likes of the post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCollection'
        "404": #post not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
        "401":
          $ref: '#/components/responses/Unauthorized'
...
  /users/{userId}/posts/{postId}/comments/{commentId}/likes:
    description: This endpoint handles the collection of likes of a comment.
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/postId'
      - $ref: '#/components/parameters/commentId'

    get:
      tags: ["like"]
      operationId: getCommentLikes
      summary: Get all the likes of a comment
      description: |
        This request is used to get all the likes of a comment.
        The postId is passed as a path parameter.
        The response will retun the list of likes of the post.
      responses:
        "200":
          description: The list of likes of the comment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCollection'
        "404": #post not found
          $ref: '#/components/responses/NotFound'
        "500": #server error
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/posts/{postId}/comments/{commentId}/likes/{likeId}:
    description: A like to a comment
    parameters:
      - $ref: '#/components/parameters/userId'
      - $ref: '#/components/parameters/postId'
      - $ref: '#/components/parameters/commentId'
      - $ref: '#/components/parameters/likeId'

    put:
      tags: ["like"]
      operationId: likeComment
      summary: Like a comment
      description: |
        This request is used to like a comment.
        The postId is taken from the path
        The userId of the user who is liking is taken from the bearer token
      responses:
        "202":
          $ref: '#/components/responses/Ok'
        "404":
          $ref: '#/components/responses/NotFound'
        "500":
          $ref: '#/components/responses/InternalServerError'
        "401":
          $ref: '#/components/responses/Unauthorized'
    
    delete:
      tags: ["like"]
      operationId: unlikeComment
      summary: Unlike a comment
      description: |
        This request is used to remove a like from a comment.
        The postId is taken from the path
        The userId of the like to remove is taken from the bearer token
        The response will return the new list of likes of the comment
      responses:
        '200':
          description: The new list of users that liked the post
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/userCollection'
        '400':
          $ref: '#/components/responses/BadRequest'
        "401":
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
security:
  - bearerAuth: [] 